+++ src/op/iter_all_weighted_edges.rs
@@ -8,9 +8,9 @@
 //! let graph = vec![(0, 1, 2), (1, 2, 3), (2, 0, 4)];
 //! let mut iter = graph.iter_all_weighted_edges();
 //!
-//! assert_eq!(iter.next(), Some(&(0, 1, 2)));
-//! assert_eq!(iter.next(), Some(&(1, 2, 3)));
-//! assert_eq!(iter.next(), Some(&(2, 0, 4)));
+//! assert_eq!(iter.next(), Some((0, 1, &2)));
+//! assert_eq!(iter.next(), Some((1, 2, &3)));
+//! assert_eq!(iter.next(), Some((2, 0, &4)));
 //! assert_eq!(iter.next(), None);
 //! ```
 
@@ -34,11 +34,11 @@ use {
 /// }
 ///
 /// impl IterAllWeightedEdges<usize> for Graph {
-///     fn iter_all_weighted_edges<'a>(&'a self) -> impl Iterator<Item = &(usize, usize, usize)>
+///     fn iter_all_weighted_edges<'a>(&'a self) -> impl Iterator<Item = (usize, usize, &'a usize)>
 ///     where
 ///         usize: 'a,
 ///     {
-///         self.edges.iter()
+///         self.edges.iter().map(|(s, t, w)| (*s, *t, w))
 ///     }
 /// }
 /// ```
@@ -51,14 +51,14 @@ use {
 /// let graph = vec![(0, 1, 2), (1, 2, 3), (2, 0, 4)];
 /// let mut iter = graph.iter_all_weighted_edges();
 ///
-/// assert_eq!(iter.next(), Some(&(0, 1, 2)));
-/// assert_eq!(iter.next(), Some(&(1, 2, 3)));
-/// assert_eq!(iter.next(), Some(&(2, 0, 4)));
+/// assert_eq!(iter.next(), Some((0, 1, &2)));
+/// assert_eq!(iter.next(), Some((1, 2, &3)));
+/// assert_eq!(iter.next(), Some((2, 0, &4)));
 /// assert_eq!(iter.next(), None);
 /// ```
 pub trait IterAllWeightedEdges<W> {
     /// Returns an iterator that iterates over all weighted edges in a graph.
-    fn iter_all_weighted_edges<'a>(&'a self) -> impl Iterator<Item = &'a (usize, usize, W)>
+    fn iter_all_weighted_edges<'a>(&'a self) -> impl Iterator<Item = (usize, usize, &'a W)>
     where
         W: 'a;
 }
@@ -66,11 +66,11 @@ pub trait IterAllWeightedEdges<W> {
 // Slice
 
 impl<W> IterAllWeightedEdges<W> for [(usize, usize, W)] {
-    fn iter_all_weighted_edges<'a>(&'a self) -> impl Iterator<Item = &'a (usize, usize, W)>
+    fn iter_all_weighted_edges<'a>(&'a self) -> impl Iterator<Item = (usize, usize, &'a W)>
     where
         W: 'a,
     {
-        self.iter()
+        self.iter().map(|(s, t, w)| (*s, *t, w))
     }
 }
 
@@ -80,11 +80,11 @@ impl<W, H> IterAllWeightedEdges<W> for HashSet<(usize, usize, W), H>
 where
     H: BuildHasher,
 {
-    fn iter_all_weighted_edges<'a>(&'a self) -> impl Iterator<Item = &'a (usize, usize, W)>
+    fn iter_all_weighted_edges<'a>(&'a self) -> impl Iterator<Item = (usize, usize, &'a W)>
     where
         W: 'a,
     {
-        self.iter()
+        self.iter().map(|(s, t, w)| (*s, *t, w))
     }
 }
 
@@ -101,9 +101,9 @@ mod tests {
         let graph = vec![(0, 1, 2), (1, 2, 3), (2, 0, 4)];
         let mut iter = graph.iter_all_weighted_edges();
 
-        assert_eq!(iter.next(), Some(&(0, 1, 2)));
-        assert_eq!(iter.next(), Some(&(1, 2, 3)));
-        assert_eq!(iter.next(), Some(&(2, 0, 4)));
+        assert_eq!(iter.next(), Some((0, 1, &2)));
+        assert_eq!(iter.next(), Some((1, 2, &3)));
+        assert_eq!(iter.next(), Some((2, 0, &4)));
         assert_eq!(iter.next(), None);
     }
 
@@ -112,9 +112,9 @@ mod tests {
         let graph: &[(usize, usize, usize)] = &[(0, 1, 2), (1, 2, 3), (2, 0, 4)];
         let mut iter = graph.iter_all_weighted_edges();
 
-        assert_eq!(iter.next(), Some(&(0, 1, 2)));
-        assert_eq!(iter.next(), Some(&(1, 2, 3)));
-        assert_eq!(iter.next(), Some(&(2, 0, 4)));
+        assert_eq!(iter.next(), Some((0, 1, &2)));
+        assert_eq!(iter.next(), Some((1, 2, &3)));
+        assert_eq!(iter.next(), Some((2, 0, &4)));
         assert_eq!(iter.next(), None);
     }
 
@@ -123,9 +123,9 @@ mod tests {
         let graph = [(0, 1, 2), (1, 2, 3), (2, 0, 4)];
         let mut iter = graph.iter_all_weighted_edges();
 
-        assert_eq!(iter.next(), Some(&(0, 1, 2)));
-        assert_eq!(iter.next(), Some(&(1, 2, 3)));
-        assert_eq!(iter.next(), Some(&(2, 0, 4)));
+        assert_eq!(iter.next(), Some((0, 1, &2)));
+        assert_eq!(iter.next(), Some((1, 2, &3)));
+        assert_eq!(iter.next(), Some((2, 0, &4)));
         assert_eq!(iter.next(), None);
     }
 
@@ -136,9 +136,9 @@ mod tests {
 
         let mut iter = graph.iter_all_weighted_edges();
 
-        assert_matches!(iter.next(), Some(&(0, 1, 2) | &(1, 2, 3) | &(2, 0, 4)));
-        assert_matches!(iter.next(), Some(&(0, 1, 2) | &(1, 2, 3) | &(2, 0, 4)));
-        assert_matches!(iter.next(), Some(&(0, 1, 2) | &(1, 2, 3) | &(2, 0, 4)));
+        assert_matches!(iter.next(), Some((0, 1, &2) | (1, 2, &3) | (2, 0, &4)));
+        assert_matches!(iter.next(), Some((0, 1, &2) | (1, 2, &3) | (2, 0, &4)));
+        assert_matches!(iter.next(), Some((0, 1, &2) | (1, 2, &3) | (2, 0, &4)));
         assert_matches!(iter.next(), None);
     }
 }